Combine receiver classes into a single file.
